SERVER/requirements.txt
whisper
pyannote.audio
pygame
gradio
fastapi
uvicorn
peft
bitsandbytes
accelerate
datasets
transformers
datasets
piper-tts
xttsv2
gtts
TTS
huggingface_hub
requests
